Include root module when applying common settings
The version update plugin needs to be able to resolve dependencies from
Maven Central but it runs in the context of the root project which
didn't have that enabled. This changes includes the root project when
applying common settings including mavenCentral enablement, plus group
and version attribute setting.

There wasn't a great reason for applying settings to every project
except the root other than it didn't strictly need them.